requirements.txt
defusedxml==0.6.0
diff-match-patch==20181111
#psycopg2>=2.7,<3.0
#psycopg2-binary
Django==2.2.6
django-cors-headers==3.1.1
django-filter==2.2.0
django-import-export==1.2.0
djangorestframework==3.10.3
djangorestframework-filters==0.3.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.0
PyJWT==1.7.1
python-decouple==3.1
pytz==2019.3
PyYAML==5.1.2
sqlparse==0.3.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0
gunicorn==19.7.1
whitenoise==3.3.1
